requirements_prod.txt
google-genai
google-adk
streamlit
python-dotenv
arxiv
duckduckgo-search
beautifulsoup4
requests
pydantic
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
numpy
scipy
